made spellCheck class
